clean tests and add e2ee test
