+++ packages/matrix-crdt/src/benchmark/matrix.bench.ts
@@ -45,10 +45,10 @@ async function readRoom(roomName: string) {
 }
 
 async function loadTest() {
-  const setup = await createRandomMatrixClientAndRoom(
-    "public-read-write",
-    false
-  );
+  const setup = await createRandomMatrixClientAndRoom({
+    permissions: "public-read-write",
+    encrypted: false,
+  });
 
   await setRoomContents(setup.client, setup.roomId);
   const server = await createSimpleServer(() => readRoom(setup.roomName));
@@ -66,3 +66,134 @@ async function loadTest() {
 //   await loadTest();
 // }, 30000);
 loadTest();
+
+/*
+class TestReader {
+  private static CREATED = 0;
+
+  public messages: any[] | undefined;
+  private reader: MatrixReader | undefined;
+  constructor(
+    private config: any,
+    private roomId: string,
+    private client?: MatrixClient
+  ) {}
+
+  async initialize() {
+    const guestClient =
+      this.client || (await createMatrixGuestClient(matrixTestConfig));
+    this.reader = new MatrixReader(
+      guestClient,
+      this.roomId,
+      new MatrixCRDTEventTranslator()
+    );
+
+    this.messages = await this.reader.getInitialDocumentUpdateEvents();
+    console.log("created", TestReader.CREATED++);
+    this.reader.onEvents((msgs) => {
+      // console.log("on message");
+      this.messages!.push.apply(this.messages, msgs.events);
+    });
+    this.reader.startPolling();
+  }
+
+  dispose() {
+    this.reader?.dispose();
+  }
+}
+
+// Breaks at 500 parallel requests locally
+it.skip("handles parallel live messages", async () => {
+  const PARALLEL = 500;
+  let messageId = 0;
+  const setup = await createRandomMatrixClientAndRoom({
+    permissions: "public-read",
+    encrypted: false,
+  });
+
+  const readers = [];
+  try {
+    const client = await createMatrixGuestClient(matrixTestConfig);
+    for (let i = 0; i < PARALLEL; i++) {
+      // const worker = new Worker(__dirname + "/worker.js", {
+      //   workerData: {
+      //     path: "./MatrixReader.test.ts",
+      //   },
+      // });
+      readers.push(new TestReader(matrixTestConfig, setup.roomId, client));
+    }
+
+    // return;
+    await Promise.all(readers.map((reader) => reader.initialize()));
+
+    while (messageId < 10) {
+      // console.log("send message");
+      await sendMessage(setup.client, setup.roomId, "message " + ++messageId);
+    }
+
+    await new Promise((resolve) => setTimeout(resolve, 10000));
+    readers.map((r) => validateMessages(r.messages!, messageId));
+  } finally {
+    readers.map((r) => r.dispose());
+  }
+});
+
+// gets slow at around 500 messages, but calls to http://localhost:8888/_matrix/static/ also at around 1k
+it.skip("handles parallel live messages autocannon", async () => {
+  const PARALLEL = 500;
+
+  let messageId = 0;
+  const setup = await createRandomMatrixClientAndRoom({
+    permissions: "public-read",
+    encrypted: false,
+  });
+
+  const client = await createMatrixGuestClient(matrixTestConfig);
+  const reader = new MatrixReader(
+    client,
+    setup.roomId,
+    new MatrixCRDTEventTranslator()
+  );
+  try {
+    await reader.getInitialDocumentUpdateEvents();
+
+    const params = {
+      access_token: client.http.opts.accessToken,
+      from: reader.latestToken,
+      room_id: setup.roomId,
+      timeout: 30000,
+    };
+
+    // send some new messages beforehand
+    while (messageId < 10) {
+      // console.log("send message");
+      await sendMessage(setup.client, setup.roomId, "message " + ++messageId);
+    }
+
+    const uri =
+      "http://" +
+      HOMESERVER_NAME +
+      "/_matrix/client/r0/events?" +
+      qs.stringify(params);
+    autocannonSeparateProcess([
+      "-c",
+      PARALLEL + "",
+      "-a",
+      PARALLEL + "",
+      "-t",
+      "20",
+      uri,
+    ]);
+
+    // send some new messages in parallel / after
+    while (messageId < 20) {
+      // console.log("send message");
+      await sendMessage(setup.client, setup.roomId, "message " + ++messageId);
+    }
+
+    await new Promise((resolve) => setTimeout(resolve, 10000));
+  } finally {
+    reader.dispose();
+  }
+});
+*/
