+++ packages/matrix-crdt/src/memberReader/MatrixMemberReader.test.ts
@@ -19,11 +19,15 @@ beforeAll(async () => {
   await ensureMatrixIsRunning();
 });
 
+/**
+ * This test validates whether MatrixMemberReader correctly detects write access
+ * when users join a room
+ */
 it("handles room joins", async () => {
-  const setupA = await createRandomMatrixClientAndRoom(
-    "public-read-write",
-    false
-  );
+  const setupA = await createRandomMatrixClientAndRoom({
+    permissions: "public-read-write",
+    encrypted: false,
+  });
   const userB = await createRandomMatrixClient();
   const guestClient = await createMatrixGuestClient(matrixTestConfig);
 
@@ -37,6 +41,28 @@ it("handles room joins", async () => {
   await readerC.startPolling();
   await memberC.initialize();
 
+  // alternative, now we use startClient, we can sync to something like this:
+
+  // const room = setupA.client.getRoom(setupA.roomId)!;
+  // expect(
+  //   room.currentState.maySendEvent(
+  //     "m.room.message",
+  //     guestClient.credentials.userId!
+  //   )
+  // ).toBe(false);
+  // expect(
+  //   room.currentState.maySendEvent(
+  //     "m.room.message",
+  //     setupA.client.credentials.userId!
+  //   )
+  // ).toBe(true);
+  // expect(
+  //   room.currentState.maySendEvent(
+  //     "m.room.message",
+  //     userB.client.credentials.userId!
+  //   )
+  // ).toBe(false);
+
   expect(memberC.hasWriteAccess(guestClient.credentials.userId!)).toBe(false);
   expect(memberC.hasWriteAccess(setupA.client.credentials.userId!)).toBe(true);
   expect(memberC.hasWriteAccess(userB.client.credentials.userId!)).toBe(false);
@@ -49,11 +75,15 @@ it("handles room joins", async () => {
   readerC.dispose();
 }, 30000);
 
+/**
+ * This test checks whether MatrixMemberReader correctly checks
+ * power level events to see if a user has write access or not
+ */
 it("handles room power levels", async () => {
-  const setupA = await createRandomMatrixClientAndRoom(
-    "public-read-write",
-    false
-  );
+  const setupA = await createRandomMatrixClientAndRoom({
+    permissions: "public-read-write",
+    encrypted: false,
+  });
   const userB = await createRandomMatrixClient();
   const guestClient = await createMatrixGuestClient(matrixTestConfig);
